+++ lib/byob/application.ex
@@ -9,6 +9,7 @@ defmodule Byob.Application do
   def start(_type, _args) do
     :ets.new(:youtube_comments_cache, [:named_table, :public, :set])
     :ets.new(:youtube_videos_cache, [:named_table, :public, :set])
+    :ets.new(:youtube_playlists_cache, [:named_table, :public, :set])
 
     children = [
       ByobWeb.Telemetry,